BackEnd/src/requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.2.1
exceptiongroup==1.3.0
Flask==3.0.3
flask-cors==6.0.1
Flask-RESTful==0.3.10
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-genai==1.22.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
moviepy==2.2.1
numpy==2.2.6
pillow==11.2.1
proglog==0.1.12
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dotenv==1.1.1
pytube==15.0.0
pytubefix==9.2.2
pytz==2025.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==15.0.1
Werkzeug==3.0.3
